services/backend/requirements.txt
fastapi>=0.115,<1
uvicorn>=0.32,<1
psycopg[binary]>=3.2,<4
pydantic>=2,<3
python-dotenv>=1.0,<2
scipy>=1.11,<2
pytest>=8,<9
httpx>=0.27,<1
